Refactor naming to be associated with the origin of activity
